refactor: Add cancellationToken support, clean up namespace, fix inheritance, and remove redundant Id field
